requirements.txt
pandas
GitPython
tk
regex
python-time
